Adding pytest unittests.
